+++ src/runtime/test/test_ConfigsMerger.cpp
@@ -40,29 +40,84 @@ TEST(ConfigsMergerTestGroup, workdir) {
 }
 
 TEST(ConfigsMergerTestGroup, environment) {
+    namespace rj = rapidjson;
     auto configRAII = test_utility::config::makeConfig();
     auto& config = configRAII.config;
+    auto& allocator = config->json.GetAllocator();
     auto metadata = common::ImageMetadata{};
 
-    // only host environment
+    rj::Value environmentValue(rj::kObjectType);
+
+    // empty environment
+    {
+        config->commandRun.hostEnvironment = {};
+        metadata.env = {};
+        CHECK((ConfigsMerger{config, metadata}.getEnvironmentInContainer().empty()));
+    }
+    // no image metadata variables
     {
         config->commandRun.hostEnvironment = {{"KEY", "HOST_VALUE"}};
         metadata.env = {};
         auto expectedEnvironment = std::unordered_map<std::string, std::string>{{"KEY", "HOST_VALUE"}};
         CHECK((ConfigsMerger{config, metadata}.getEnvironmentInContainer() == expectedEnvironment));
     }
-    // only metadata environment
+    // no host variables
     {
         config->commandRun.hostEnvironment = {};
-        metadata.env = {{"KEY", "CONTAINER_VALUE"}};
-        auto expectedEnvironment = std::unordered_map<std::string, std::string>{{"KEY", "CONTAINER_VALUE"}};
+        metadata.env = {{"KEY", "IMAGE_VALUE"}};
+        auto expectedEnvironment = std::unordered_map<std::string, std::string>{{"KEY", "IMAGE_VALUE"}};
         CHECK((ConfigsMerger{config, metadata}.getEnvironmentInContainer() == expectedEnvironment));
     }
-    // metadata environment overrides host environment
+    // image metadata environment overrides host environment
     {
         config->commandRun.hostEnvironment = {{"KEY", "HOST_VALUE"}};
-        metadata.env = {{"KEY", "CONTAINER_VALUE"}};
-        auto expectedEnvironment = std::unordered_map<std::string, std::string>{{"KEY", "CONTAINER_VALUE"}};
+        metadata.env = {{"KEY", "IMAGE_VALUE"}};
+        auto expectedEnvironment = std::unordered_map<std::string, std::string>{{"KEY", "IMAGE_VALUE"}};
+        CHECK((ConfigsMerger{config, metadata}.getEnvironmentInContainer() == expectedEnvironment));
+    }
+    // config file set environment overrides both host and image metadata
+    {
+        rj::Value setValue(rj::kObjectType);
+        setValue.AddMember("SARUS_CONFIG_SET",
+                           rj::Value{"config_set_value", allocator},
+                           allocator);
+        environmentValue.AddMember("set", setValue, allocator);
+        config->json.AddMember("environment", environmentValue, allocator);
+
+        config->commandRun.hostEnvironment = {{"SARUS_CONFIG_SET", "HOST_VALUE"}};
+        metadata.env = {{"SARUS_CONFIG_SET", "IMAGE_VALUE"}};
+        auto expectedEnvironment = std::unordered_map<std::string, std::string>{{"SARUS_CONFIG_SET", "config_set_value"}};
+        CHECK((ConfigsMerger{config, metadata}.getEnvironmentInContainer() == expectedEnvironment));
+    }
+    // config file prepend/append
+    {
+        rj::Value prependValue(rj::kObjectType);
+        prependValue.AddMember("SARUS_CONFIG_PREPEND_APPEND",
+                               rj::Value{"config_prepend_value", allocator},
+                               allocator);
+        rj::Value appendValue(rj::kObjectType);
+        appendValue.AddMember("SARUS_CONFIG_PREPEND_APPEND",
+                              rj::Value{"config_append_value", allocator},
+                              allocator);
+        config->json["environment"].AddMember("prepend", prependValue, allocator);
+        config->json["environment"].AddMember("append", appendValue, allocator);
+
+        config->commandRun.hostEnvironment = {};
+        metadata.env = {{"SARUS_CONFIG_PREPEND_APPEND", "IMAGE_VALUE"}};
+        auto expectedEnvironment = std::unordered_map<std::string, std::string>{
+            {"SARUS_CONFIG_SET", "config_set_value"}, {"SARUS_CONFIG_PREPEND_APPEND", "config_prepend_value:IMAGE_VALUE:config_append_value"}};
+        CHECK((ConfigsMerger{config, metadata}.getEnvironmentInContainer() == expectedEnvironment));
+    }
+    // config file unset
+    {
+        rj::Value unsetValue(rj::kArrayType);
+        unsetValue.PushBack("SARUS_CONFIG_UNSET", allocator);
+        config->json["environment"].AddMember("unset", unsetValue, allocator);
+
+        config->commandRun.hostEnvironment = {};
+        metadata.env = {{"SARUS_CONFIG_UNSET", "IMAGE_VALUE"}};
+        auto expectedEnvironment = std::unordered_map<std::string, std::string>{
+            {"SARUS_CONFIG_SET", "config_set_value"}, {"SARUS_CONFIG_PREPEND_APPEND", "config_prepend_value:config_append_value"}};
         CHECK((ConfigsMerger{config, metadata}.getEnvironmentInContainer() == expectedEnvironment));
     }
 }